+++ docs/pages/visual-test/Link-dark.js
@@ -1,17 +1,32 @@
 // @flow strict
 import { type Node } from 'react';
-import { Text, Link, Box, ColorSchemeProvider } from 'gestalt';
+import { Text, Link, Box, ColorSchemeProvider, Flex } from 'gestalt';
 
 export default function Screenshot(): Node {
   return (
     <ColorSchemeProvider colorScheme="dark">
       <Box color="white" display="inlineBlock" padding={4}>
-        <Text inline>
-          Visit our{' '}
-          <Link inline href="https://pinterest.com">
-            Business Help Center
-          </Link>
-        </Text>
+        <Flex direction="column" gap={2}>
+          ejefcbcvhultvfdhebvtulclubiljucrrkhhljgkditb
+          <Text inline>
+            Visit our{' '}
+            <Link inline href="https://pinterest.com">
+              Business Help Center
+            </Link>
+          </Text>
+          <Text inline size="400" color="success">
+            <Link
+              href="https://authy.com/download/"
+              inline
+              externalLinkIcon={{ size: '400', color: 'success' }}
+              target="blank"
+              rel="nofollow"
+            >
+              MyBusiness.com
+            </Link>{' '}
+            was successfully claimed
+          </Text>
+        </Flex>
       </Box>
     </ColorSchemeProvider>
   );