+++ docs/pages/visual-test/Link.js
@@ -1,16 +1,30 @@
 // @flow strict
 import { type Node } from 'react';
-import { Text, Link, Box } from 'gestalt';
+import { Text, Link, Box, Flex } from 'gestalt';
 
 export default function Screenshot(): Node {
   return (
     <Box color="white" display="inlineBlock" padding={4}>
-      <Text inline>
-        Visit our{' '}
-        <Link inline href="https://pinterest.com">
-          Business Help Center
-        </Link>
-      </Text>
+      <Flex direction="column" gap={2}>
+        <Text inline>
+          Visit our{' '}
+          <Link inline href="https://pinterest.com">
+            Business Help Center
+          </Link>
+        </Text>
+        <Text inline size="400" color="success">
+          <Link
+            href="https://authy.com/download/"
+            inline
+            externalLinkIcon={{ size: '400', color: 'success' }}
+            target="blank"
+            rel="nofollow"
+          >
+            MyBusiness.com
+          </Link>{' '}
+          was successfully claimed
+        </Text>
+      </Flex>
     </Box>
   );
 }